Fixed Attacking Bug near ledge
When enemies were on a ledge and attacked, they'd enter the chase state once the player leaves but they'd get stuck in the last frame of the attack animation because there was only logic in chase enter to go into chase but because there is a ledge they couldn't do that. Now implemented logic to go into idle and enemies no longer get stuck in attack.